# Docs/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AUSTA Cockpit API
  description: |
    Medical Audit Platform API with AI-powered decision support.
    
    This API provides endpoints for managing medical authorization cases, 
    AI-assisted analysis, audit decisions, and comprehensive analytics.
    
    ## Authentication
    
    The API uses JWT tokens for authentication. Obtain a token through the 
    `/auth/login` endpoint and include it in the Authorization header:
    
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    
    API calls are rate limited to ensure fair usage:
    - Anonymous: 100 requests/hour
    - Authenticated: 1000 requests/hour
    - Premium: 10000 requests/hour
    
    ## Environments
    
    - Development: `http://localhost:3000/api/v1`
    - Staging: `https://staging-api.austa.com.br/v1`
    - Production: `https://api.austa.com.br/v1`
    
  version: 1.0.0
  contact:
    name: AUSTA Support
    email: suporte@austa.com.br
    url: https://docs.austa.com.br
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://staging-api.austa.com.br/v1
    description: Staging server
  - url: https://api.austa.com.br/v1
    description: Production server
...
paths:
...
  /analytics/dashboard:
    get:
      tags:
        - Analytics
      summary: Get dashboard metrics
      description: Retrieve key performance metrics for the dashboard
      operationId: getDashboardMetrics
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          description: Time period for metrics
          schema:
            type: string
            enum: [today, week, month, quarter, year]
            default: month
        - name: auditorId
          in: query
          description: Filter by specific auditor
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Metrics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardMetrics'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
...
  schemas:
...
    DashboardMetrics:
      type: object
      properties:
        summary:
          type: object
          properties:
            totalCases:
              type: integer
            pendingCases:
              type: integer
            completedCases:
              type: integer
            averageProcessingTime:
              type: number
              description: Average time in minutes
            approvalRate:
              type: number
              format: float
            aiAccuracy:
              type: number
              format: float
        casesByStatus:
          type: array
          items:
            type: object
            properties:
              status:
                type: string
              count:
                type: integer
              percentage:
                type: number
                format: float
        casesByPriority:
          type: array
          items:
            type: object
            properties:
              priority:
                type: string
              count:
                type: integer
        performanceTrend:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              casesProcessed:
                type: integer
              averageTime:
                type: number
              approvalRate:
                type: number
                format: float
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
            timestamp:
              type: string
              format: date-time
            path:
              type: string
            requestId:
              type: string
              format: uuid
...
  responses:
...
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...